+++ modules/ismp/pallet/Cargo.toml
@@ -19,18 +19,16 @@ sp-api = { workspace = true }
 merkle-mountain-range = { workspace = true }
 pallet-timestamp = { workspace = true, optional = true }
 sp-consensus-aura = { workspace = true }
- # polytope labs
 ismp = { workspace = true }
-fortuples = "0.9.1"
+codec = { workspace = true }
+scale-info = { workspace = true }
 
 # crates.io
-codec = { package = "parity-scale-codec", version = "3.1.3", default-features = false }
+fortuples = "0.9.1"
 log = { version = "0.4.17", default-features = false }
-scale-info = { version = "2.1.1", default-features = false, features = ["derive"] }
 serde = { version = "1.0.136", features = ["derive"], optional = true }
 derive_more = { version = "0.99.17", default-features = false, features = ["from", "into", "display"] }
 enum-as-inner = "=0.5.1"
-
 
 [dev-dependencies]
 env_logger = "0.10.0"
