+++ modules/ismp/pallet/fishermen/Cargo.toml
@@ -13,14 +13,12 @@ sp-runtime = { workspace = true }
 sp-core = { workspace = true }
 sp-std = { workspace = true }
 sp-io = { workspace = true }
+scale-info = { workspace = true }
+codec = { workspace = true }
 
 # polytope labs
 ismp = { workspace = true }
 pallet-ismp = { workspace = true }
-
-scale-info = { version = "2.1.1", default-features = false, features = ["derive"] }
-codec = { package = "parity-scale-codec", version = "3.1.3", default-features = false }
-
 
 [features]
 default = ["std"]
