+++ modules/ismp/pallet/src/mmr/utils.rs
@@ -13,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::mmr_primitives::{LeafIndex, NodeIndex};
+use crate::mmr::primitives::{LeafIndex, NodeIndex};
 
 /// MMR nodes & size -related utilities.
 pub struct NodesUtils {
