+++ modules/ismp/pallet/testsuite/src/lib.rs
@@ -1,0 +1,2 @@
+mod runtime;
+mod tests;
