+++ modules/ismp/pallet/testsuite/src/tests/mod.rs
@@ -1,0 +1,3 @@
+mod pallet_fishermen;
+mod pallet_ismp;
+mod pallet_ismp_relayer;
