+++ parachain/runtimes/gargantua/src/ismp.rs
@@ -81,7 +81,7 @@ impl pallet_ismp_relayer::Config for Runtime {
     type RuntimeEvent = RuntimeEvent;
 }
 
-impl ismp_host_executive::Config for Runtime {}
+impl pallet_ismp_host_executive::Config for Runtime {}
 
 impl IsmpModule for ProxyModule {
     fn on_accept(&self, request: Post) -> Result<(), Error> {
